+++ CHANGELOG.md
@@ -13,6 +13,7 @@ with nodejs and emscripten)
 - Added preliminary support for ESP32P4 (no networking support yet).
 - Added memory info in `out_of_memory` crash logs to help developers fix memory issues.
 - Added documentation and function specs for uart driver
+- Added `uart:read/2` with a timeout parameter.
 
 ### Fixed